NS-28-08-2020

+++ app/src/main/java/com/naposystems/napoleonchat/utility/UploadResult.kt
@@ -15,9 +15,9 @@ sealed class UploadResult {
         val cause: Exception? = null
     ) : UploadResult()
 
-    data class CompressProgress(val attachment: Attachment, val progress: Long, val job: ProducerScope<UploadResult>) : UploadResult()
+    data class CompressProgress(val attachment: Attachment, val progress: Float, val job: ProducerScope<UploadResult>) : UploadResult()
 
-    data class Progress(val attachment: Attachment, val progress: Long, val job: ProducerScope<UploadResult>) : UploadResult()
+    data class Progress(val attachment: Attachment, val progress: Float, val job: ProducerScope<UploadResult>) : UploadResult()
 
     data class Complete(val attachment: Attachment) : UploadResult()
 }
